requirements.txt
aiohttp==3.9.1
requests==2.31.0
firebase-admin==6.3.0
google-cloud-firestore==2.14.0
google-auth==2.25.2
google-api-python-client==2.110.0
cryptography==41.0.7
Pillow==10.1.0
psutil==5.9.6
joblib==1.5.2
flask==3.1.2
pandas==2.3.3